PoC: Make sure the new window replaces the current
This is not at all ready to be merged, but is an experiment trying to
implement #1.

By converting the original window into a split, the newly-created window
is added to that split (which guarantees that they're adjacent). Since
the original window is then hidden, the new window effectively takes its
place.

This has a few caveats:

- There's a "wait 1s" hardcoded there:
  - If the new window is faster, this is weird and glitchy.
  - If the new window is slower, this doesn't work.
- When the new window _exits_, the swallowed window won't always take
  its place: it'll be placed relative to the last-focused window, and
  I can't think of any way around this other than continuously
  monitoring the tree to find our right position.

The timeout can we worked around by waiting for _any_ window to appear,
and assuming the first new window to spawn is the one we're creating.
This might be flaky, but I suspect it'd work 90% of the time.

An alternative implementation could also be using a new stacked/tabbed
layout, where the new window is a second in that container. If sway
allowed us to hide the stack/tab bar on that container, we'd be 100%
covered.