Avoid the long wait when running a command
And actually wait until a new window appears. This can be problematic if
an UNRELATED window appears in the meantime, however, the Event object
DOES have the `pid` of the process, and since it's our child, we should
be able to compare the `pid` and confirm if this is the right window or
not.

Sadly, my Rust isn't good enough to unwrap the event and extract the
pid.